pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lectio-sync"
version = "0.1.0"
description = "Parse Lectio Advanced Schedule HTML to iCalendar (.ics)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
  "python-dateutil==2.9.0.post0",
]

[project.scripts]
lectio-ics = "lectio_sync.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
